moved util functions to 1 file
